feat(sdk): add jira posture helper modules
