+++ sdk/python/cerebro_sdk/__init__.py
@@ -1,7 +1,31 @@
 from .client import APIError, Client, IntegrationClient
+from .jira import (
+    JiraAdminPosture,
+    JiraMarketplaceAppPosture,
+    JiraPostureFinding,
+    JiraProjectPosture,
+    JiraWorkspaceGraphLayering,
+    JiraWorkspacePosture,
+    OnboardJiraWorkspacePostureResult,
+    build_jira_posture_findings,
+    build_jira_workspace_claims,
+    load_jira_workspace_graph_layering,
+    onboard_jira_workspace_posture,
+)
 
 __all__ = [
     "APIError",
     "Client",
     "IntegrationClient",
+    "JiraAdminPosture",
+    "JiraMarketplaceAppPosture",
+    "JiraPostureFinding",
+    "JiraProjectPosture",
+    "JiraWorkspaceGraphLayering",
+    "JiraWorkspacePosture",
+    "OnboardJiraWorkspacePostureResult",
+    "build_jira_posture_findings",
+    "build_jira_workspace_claims",
+    "load_jira_workspace_graph_layering",
+    "onboard_jira_workspace_posture",
 ]
