+++ sdk/typescript/package.json
@@ -4,6 +4,7 @@
   "type": "module",
   "main": "./src/index.ts",
   "exports": {
-    ".": "./src/index.ts"
+    ".": "./src/index.ts",
+    "./jira": "./src/jira.ts"
   }
 }
